Extract validation issue construction
